Move SourceRevision and ApplicationVersion to controller-api
